Add UserKey and UserFormat to generic OAuth
This adds `UserKey` and `UserFormat` to the configuration for the
generic OAuth provider. `UserKey` specifies the name of the key in the
body of the UserURL that contains the username. It defaults to email.

`UserFormat` optionally contains an Sprintf format to pass the username
through.

I added these in order to use traefik-forward-auth with Mastodon as an
OAuth provider.  I am using `/api/v1/accounts/verify_credentials` as the
UserURL, but it does not contain an `email` key; instead, it contains a
`username` key.